Fix: - Card List Item Home
New:
- Seller API
- Seller DI
- Seller Poduct Response
- Item Produck Adapter
- Daftar Jual ViewModel

Update:
- Item Produk UI View
- Button Logic for Adding Item

+++ app/src/main/java/binar/and3/kelompok1/secondhand/data/api/auth/SignUpResponse.kt
@@ -7,7 +7,7 @@ data class SignUpResponse(
     @SerializedName("full_name"    ) var fullName    : String? = null,
     @SerializedName("email"        ) var email       : String? = null,
     @SerializedName("password"     ) var password    : String? = null,
-    @SerializedName("phone_number" ) var phoneNumber : Int?    = null,
+    @SerializedName("phone_number" ) var phoneNumber : Long?   = null,
     @SerializedName("address"      ) var address     : String? = null,
     @SerializedName("city"         ) var city        : String? = null,
     @SerializedName("image_url"    ) var imageUrl    : String? = null
